# Opt performance of async log